Arc 1558 push app id to queue msg (#1446)
* ARC-1558 Add interface contract for message queue payload

* Extrac the message payload typings into one file
