+++ src/github/push.test.ts
@@ -7,7 +7,8 @@ import { Installation } from "models/installation";
 import { Subscription } from "models/subscription";
 import { Application } from "probot";
 import { waitUntil } from "test/utils/wait-until";
-import { pushQueueMessageHandler, PushQueueMessagePayload } from "../sqs/push";
+import { pushQueueMessageHandler } from "../sqs/push";
+import { PushQueueMessagePayload } from "../sqs/sqs.types";
 import { Context } from "../sqs/sqs";
 import { Message } from "aws-sdk/clients/sqs";
 import { sqsQueues } from "../sqs/queues";
