+++ src/sync/branches.ts
@@ -3,7 +3,7 @@ import { GitHubAPI } from "probot";
 import { Repository } from "models/subscription";
 import { GitHubInstallationClient } from "../github/client/github-installation-client";
 import Logger from "bunyan";
-import { BackfillMessagePayload } from "~/src/sqs/backfill";
+import { BackfillMessagePayload } from "~/src/sqs/sqs.types";
 import { NumberFlags } from "config/feature-flags";
 import { getCommitSinceDate } from "~/src/sync/sync-utils";
