+++ src/sync/commits.ts
@@ -7,7 +7,7 @@ import { CommitQueryNode } from "../github/client/github-queries";
 import { JiraCommitData } from "src/interfaces/jira";
 import { NumberFlags } from "config/feature-flags";
 import { TaskPayload } from "~/src/sync/installation";
-import { BackfillMessagePayload } from "~/src/sqs/backfill";
+import { BackfillMessagePayload } from "~/src/sqs/sqs.types";
 import { getCommitSinceDate } from "~/src/sync/sync-utils";
 
 const fetchCommits = async (gitHubClient: GitHubInstallationClient, repository: Repository, commitSince?: Date, cursor?: string | number, perPage?: number) => {
